pyproject.toml
[project]
name = "chirpminds"
version = "0.1.0"
description = "AI toolkit for things that chirp"
readme = "README.md"
authors = [
    { name = "Ankur Kumar", email = "ank@leoank.me" }
]
requires-python = ">=3.10"
dependencies = [
  "ultralytics",
  "fiftyone",
  "label-studio",
  "label-studio-sdk",
  "ffmpeg-python",
  "gdown",
  "supervision",
  "moviepy",
  "pyyaml>=6.0.2",
]

[project.optional-dependencies]
cpu = [
  "torch",
  "torchvision",
]
cu124 = [
  "torch",
  "torchvision",
]
dev = [
  "pytest",
  "ruff",
  "build",
  "twine",
  "jupytext",
  "ipython",
  "jupyterlab",
  "jupyterlab-quarto",
  "jupyter-collaboration",
  "jupyterlab-horizon-theme",
  "nbdev"
]

[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu124" },
  ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu124", extra = "cu124" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu124", extra = "cu124" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[project.scripts]
thursday = "thursday:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.jupytext.formats]
"notebooks/ipynb" = "ipynb"
"notebooks/pypct" = "py:percent"

[tool.pyright]
exclude = [ "src", "notebooks"]
